Preserve Unifi config across container upgrades
We want to be able to upgrade/replace the Unifi container from time to
time (for example, when new controller versions are released). To do so
without losing our entire network setup, we need to bind the
configuration directory to the host.

This also exposes the container's logs to the host, too, so we can
easily interrogate those, and ignores these directories in git.