Allow Mongo to save to Vagrant shared folders
Mongo [doesn't support saving to Virtualbox shared
folders](https://github.com/docker-library/docs/pull/470/files) (which
use the `vboxsf` file system), so we're using NFS shares instead. They
require a dedicated private network to work for reasons I don't fully
understand (but that's what Vagrant's docs say:
https://developer.hashicorp.com/vagrant/docs/synced-folders/nfs#prerequisites).

Also of note is that [Ubuntu 22.04 doesn't support NFS over
UDP](https://github.com/hashicorp/vagrant/issues/12758), so we also
have to disable that.

Finally, we need to run the Unifi container as root, otherwise it can't
write to the NFS share. I tried updating the permissions to allow it to
run as a `unifi` user, and add a `unifi`, but there's already a user
with uid 999 in our base box, and and and… So I gave up for now.