Use UniFi Poller to export stats for Prometheus
We want a way for Prometheus to collect stats about our network, so
we're using the [UniFi Poller](https://unpoller.com). It's primarily
geared toward InfluxDB and Loki (I wonder why?), but also supports
Prometheus.

Unpoller requires a user with read access to be able to query the UniFi
API, so we're also adding a (gitignored) .env file in which to store
the UniFi credentials.