Allow UniFi devices to complete adoption
Devices can't complete adoption because they can't find the controller.
We could do [the “override inform host” dance](https://github.com/jacobalberty/unifi-docker#override-inform-host-ip),
but in my experience that's a) gross and b) error prone.

This change exposes the controller on the host's network interface so
it can use the [normal layer 2 adoption
methods](https://github.com/jacobalberty/unifi-docker/blob/master/Side-Projects.md#other-techniques-for-adoption).

Because the controller is on the host network, the UniFi poller can't
reach the controller unless it's also on the host network, so this moves
the poller to the host network too and updates the URL.